Completando as mudanças
