+++ src/components/Loading.css
@@ -1,4 +1,4 @@
-.conteiner {
+.container {
   display: flex;
   flex-direction: column;
   margin: auto;
